feat(esp_idf): vendor harvested ESP-IDF v6.1 shadow headers
Replace hand-written declaration headers with sdk_harvester artifacts

(manifest 542eb37a5dc3604c, 280 generated headers, survey scope) and adapt

the facade build:

- keep hand-written exemptions (sdkconfig_base/esp_attr/freertos/

  hal spi_types/led_strip) plus esp_check/esp_idf_wink extensions

- add default shim/include/sdkconfig.h (last in include search order)

- chips esp32 soc_caps/gpio_num forward via include_next

- implement v6 unified esp_log/esp_log_va; esp_err_t gpio_uninstall

- fix include order for corpus overlays and trace/config deps

- esp_idf ctest 28/28 (native + wasm compile + 4 corpora)

Generated-by: AI

+++ wink-micro-os/frameworks/esp_idf/src/drivers/esp_i2c_legacy.c
@@ -331,7 +331,7 @@ void esp_i2c_legacy_reset(void) {
         s_cmd_links[i].entry_count = 0;
         s_cmd_links[i].write_offset = 0;
     }
-    for (int p = 0; p < SOC_HP_I2C_NUM; p++) {
+    for (int p = 0; p < (int)SOC_HP_I2C_NUM; p++) {
         s_i2c_high_period[p] = 0;
         s_i2c_low_period[p] = 0;
         s_i2c_start_setup[p] = 0;
